Updating README file for backend
Make sure that the README of your final project includes a section titled "Kanban board" that contains: a link to your Kanban board.
a link to the screenshot of the initial state of your Kanban board (you should have prepared it when you were setting up your kanban board). a statement with the final number of team members. Copy the main link to the GH repository with your project (at this moment its final version should be in the main branch).